Part 9 Complete
In the ninth part of the course you learned some of the properties of object-oriented programming.

Programmers can expect all objects to have the methods toString, hashCode and equals because each object is of type Object in addition to their own type. Inheritance and interfaces allow use to implement methods and constructors that handle many kinds of objects. This enables us to easily extend programs and it helps with solving problems with classes and objects.