JavaFX:Event Handler, ActionListner & Application launch parameters
The event handler being used depends on what kind of user interface component we attach it to.
•	EventHandler
•	ActionEvent

Button presses are handled using a class that implements the EventHandler interface. The type of the event in these cases is ActionEvent. The interface implementation specifies what is done when a user presses a button.

•	addListener
•	ChangeListener: If we want to listen to changes made to a text field character by character, then we would use the interface ChangeListener.

text.textProperty().addListener((change, oldValue, newValue)

Application's launch parameters
•	The getParameters method returns a [Parameters] object whose method getNamed can access a hash table containing key-value pairs. The key-value pairs are given to the launch method in the form --key = value.
•	We can use this to pass in parameters at launch.
•	The below example shows how we can pass in the title of the window at launch time